21 todos (#22)
* chore: further improvements:
Added a provider for editor related dialogues.
Added link-adding, image-uploading and youtube-videos-adding to the text editor menu.
Fixed issues related with handling saved article which is currently being modified or created.

* chore: checking whether certain text editor menu items can be triggered

* chore: dialogue for the addition of links

* chore: dialogue for the addition of youtube videos

* chore: directory cleaning & session checking improvement

* chore: started working on a custom extension for the image element.

* chore: made it possible to render custom react elements properly in the editor and when reading finished article.

* chore: added a basic workflow for the custom image node view

* chore: finished working on the image view node component for text editor

* chore: image node view and image article section finished

* chore: fix informative badge styling

* chore: extracted the resizable part of image node view component and placed it in a separate component so that it can be used with different node views

* chore: resizable component
Created resizable component for the future node views which may require it.
Used the resizable component for the image & youtube node views.
Updated the list of things to do and fix.

* chore: designed a way of handling actions independently for the text editor

* chore: further improvements regarding todos list
Restructered the Dialogue Provider so that it can be used universally rather than just inside the text editor.
Created a new dialogue for logging out.
Fixed the error regarding flushSync being called inside a lifecycle method upon adding an image to the article.

* chore: further improvements for the articles and profile modification
Added the preview of the profile picture before uploading on the profile page.
Fixed the problem with relative positionioning the theme switch and user sidebar trigger on the user page.
Standardized the JSON content for the articles rather than plain HTML.
Created the unexpected error page.

* chore: minor cosmetic change to the logout dialogue

* chore: text editor improvements
Added a dialogue for changing article's banner image.
Upgraded the process of changing article's visibility.

* chore: minor fixes, dependency updates & cleaning
Updated all the packages to their latest versions.
Cleaned and organized the project's directory better.
Migrated to tailwindcss v4 with shadcn.
Created validation process for the addition of links (only https) & fixed the problem with updating user's profile picture.

* chore: deployment streamlined

* chore: added change tracking for the article writing page

* chore: added the logo image and favicon

* chore: cleaning before merge